refactor(pair): make Pair more of a value type
Pairs are more of a value type so change them up to be that way. All
operations on a single pair are methods, even those that might make new
Pairs, and rename pair.New to pair.Of since New implies memory allocation,
which often not the case. Update field names since methods now use those
names, and update tests.

Signed-off-by: Carlos Castillo <cookieo9@gmail.com>